fix: Abstracts common logic of Parser
