+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/CategorizedTestsDetectorTest.java
@@ -8,15 +8,15 @@ import org.arquillian.smart.testing.TestSelection;
 import org.arquillian.smart.testing.api.SmartTesting;
 import org.arquillian.smart.testing.configuration.Configuration;
 import org.arquillian.smart.testing.configuration.ConfigurationLoader;
-import org.arquillian.smart.testing.strategies.categorized.project.FastTaggedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.FirstAndSecondCategorizedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.FirstAndSecondTaggedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.FirstCategorizedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.FirstTaggedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.NonCategorizedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.NonTaggedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.ThirdCategorizedClass;
-import org.arquillian.smart.testing.strategies.categorized.project.ThirdTaggedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.tags.FastTaggedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.categories.FirstAndSecondCategorizedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.tags.FirstAndSecondTaggedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.categories.FirstCategorizedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.tags.FirstTaggedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.categories.NonCategorizedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.tags.NonTaggedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.categories.ThirdCategorizedClass;
+import org.arquillian.smart.testing.strategies.categorized.project.tags.ThirdTaggedClass;
 import org.arquillian.smart.testing.strategies.categorized.project.categories.FirstCategory;
 import org.arquillian.smart.testing.strategies.categorized.project.categories.SecondCategory;
 import org.junit.Before;
