--- strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/FastTaggedClass.java
@@ -1,5 +1,0 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
-
-@Fast
-public class FastTaggedClass {
-}
