--- strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/NonCategorizedClass.java
@@ -1,4 +1,0 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
-
-public class NonCategorizedClass {
-}
