--- strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/NonTaggedClass.java
@@ -1,4 +1,0 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
-
-public class NonTaggedClass {
-}
