+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/categories/FirstCategorizedClass.java
@@ -1,4 +1,4 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
+package org.arquillian.smart.testing.strategies.categorized.project.categories;
 
 import org.arquillian.smart.testing.strategies.categorized.project.categories.FirstCategory;
 import org.junit.experimental.categories.Category;
