+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/categories/ThirdCategorizedClass.java
@@ -1,4 +1,4 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
+package org.arquillian.smart.testing.strategies.categorized.project.categories;
 
 import org.arquillian.smart.testing.strategies.categorized.project.categories.ThirdCategory;
 import org.junit.experimental.categories.Category;
