+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/tags/Fast.java
@@ -1,4 +1,4 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
+package org.arquillian.smart.testing.strategies.categorized.project.tags;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
