+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/tags/FastTaggedClass.java
@@ -1,0 +1,5 @@
+package org.arquillian.smart.testing.strategies.categorized.project.tags;
+
+@Fast
+public class FastTaggedClass {
+}
