+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/tags/NonTaggedClass.java
@@ -1,0 +1,4 @@
+package org.arquillian.smart.testing.strategies.categorized.project.tags;
+
+public class NonTaggedClass {
+}
