+++ strategies/categorized/src/test/java/org/arquillian/smart/testing/strategies/categorized/project/tags/ThirdTaggedClass.java
@@ -1,4 +1,4 @@
-package org.arquillian.smart.testing.strategies.categorized.project;
+package org.arquillian.smart.testing.strategies.categorized.project.tags;
 
 import org.junit.jupiter.api.Tag;
